requirements.txt
pyrogram==2.0.106
tgcrypto
ffmpeg-python
python-dotenv
